requirements.txt
torchfun>=1.0.83
torchvision>=0.6.1
opencv_python_inference_engine>=4.0.1.2
pandas>=1.0.5
imgaug>=0.4.0
requests>=2.24.0
albumentations>=0.4.5
matplotlib>=3.2.2
kornia>=0.3.2+f4f70fe
nltk>=3.5
pypesq>=1.2.4
scikit_image>=0.16.2
torch>=1.5.1
beautifulsoup4>=4.9.1
ipython>=7.16.1
numpy>=1.19.0
path.py>=12.4.0
Pillow>=7.2.0
skimage>=0.0
timm>=0.1.30
